Super greedy or cheap auction agent.
It'll never work for free (1 is the min), otherwise it always bid to win $1 per
auction. The intersting thing about it is that it looses against the
others in many cases. Usually because one of the other picks a task up
for a good bargain when this one added up too many marginal cost to us.